requirements.txt
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
oauthlib==3.2.2
pathlib==1.0.1
requests==2.31.2
requests-oauthlib==1.3.1
